Refactored the list rendering logic (again). Let's gooo
