--- src/screens/Camera/components/ImageDetectRectDraw.tsx
@@ -1,71 +1,0 @@
-import React, {useContext} from 'react';
-import {StyleSheet} from 'react-native';
-import {TouchableOpacity} from 'react-native';
-
-import {SelectedResultContext} from '@/contexts/DetectionResultContext';
-
-import {DetectionResultType} from '@/types/detectionResult';
-
-type RectRenderProps = {
-  element: DetectionResultType;
-  index: number;
-  isReliable: boolean;
-};
-
-export const RectRender = ({element, index, isReliable}: RectRenderProps) => {
-  const {selectedResult, setSelectedResult, resizeRatio} = useContext(
-    SelectedResultContext,
-  );
-
-  const hanldePress = () => {
-    setSelectedResult(prev =>
-      prev.index === index
-        ? {result: null, index: null}
-        : {result: element.object, index},
-    );
-  };
-
-  return (
-    <TouchableOpacity
-      onPress={hanldePress}
-      key={index}
-      style={[
-        styles.rect,
-        isReliable && styles.rectFade,
-        selectedResult.index === index && styles.rectWhite,
-        // YOLOv5
-        {
-          width: (element.coordinate.x1 - element.coordinate.x0) * resizeRatio,
-          height: (element.coordinate.y1 - element.coordinate.y0) * resizeRatio,
-          left: element.coordinate.x0 * resizeRatio,
-          top: element.coordinate.y0 * resizeRatio,
-        },
-        // COCO-SSD
-        // {
-        //   left: element.bbox[0] * resizeRatio,
-        //   top: element.bbox[1] * resizeRatio,
-        //   width: element.bbox[2] * resizeRatio,
-        //   height: element.bbox[3] * resizeRatio,
-        // },
-      ]}
-    />
-  );
-};
-
-const styles = StyleSheet.create({
-  rect: {
-    position: 'absolute',
-  },
-  rectFade: {
-    borderColor: 'rgba(255, 255, 255, 0.3)',
-    borderWidth: 5,
-    borderRadius: 10,
-    shadowRadius: 4,
-  },
-  rectWhite: {
-    borderColor: 'white',
-    borderWidth: 5,
-    borderRadius: 10,
-    shadowRadius: 4,
-  },
-});
